Create DiabetesRiskPreventor
Implemented a machine learning model using the scikit-learn library in Python for diabetes risk prediction. The K-Nearest Neighbors (KNN) classifier was employed with nine neighbors. The data was split into training and testing sets, and the model was trained and evaluated, achieving a certain accuracy level. The predictions, along with the corresponding symptoms and actual results, were displayed for evaluation.